% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ady.R
\docType{data}
\name{data_ady}
\alias{data_ady}
\title{Collected works of Endre Ady}
\format{
It is a \code{data.frame}, with 1 observation, 2 variables:
\describe{
\item{doc_id}{Doc id variable for easier input into quanteda}
\item{text}{The unprocessed combined works of Endre Ady. All of the documents are contained as one observation (row).}
}
}
\usage{
data_ady
}
\description{
The data contains the collected works of the famous Hungarian poet, Endre Ady. It contains all of his works, without any cleaning or preprocessing.
}
\keyword{datasets}
