% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lawprop_sample.R
\docType{data}
\name{data_lawprop_sample}
\alias{data_lawprop_sample}
\title{Hungarian law proposals 1994-2018, (small sample)}
\format{
It is a \code{data.frame}, with 600 observation, 2 variables:
\describe{
\item{tvjav_id}{Id for each law proposal. The syntax: election-cycle_proposalname}
\item{tvjav_szoveg}{The unprocessed law proposal texts}
}
}
\source{
\url{https://cap.tk.hu/en/dataoverview}
}
\usage{
data_lawprop_sample
}
\description{
The dataset contains a sample 600 Hungarian laws proposed in the 1994-2018 period. This data is used in the 10th chapter (https://tankonyv.poltextlab.com/similarity.html).
}
\keyword{datasets}
