% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lawsample_match.R
\docType{data}
\name{data_lawsample_match}
\alias{data_lawsample_match}
\title{Match table for the laws and proposals samples}
\format{
It is a \code{data.frame}, with 600 observation, 2 variables:
\describe{
\item{tv_id}{Title of the law}
\item{tvjav_id}{Id for each law proposal. The syntax: election-cycle_proposalname}
}
}
\source{
\url{https://cap.tk.hu/en/dataoverview}
}
\usage{
data_lawsample_match
}
\description{
The dataset contains the key pairs to match data_lawprop_sample and data_lawtext_sample. This data is used in the 10th chapter (https://tankonyv.poltextlab.com/similarity.html).
}
\keyword{datasets}
