% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lawtext_1998_2006.R
\docType{data}
\name{data_lawtext_1998_2006}
\alias{data_lawtext_1998_2006}
\title{Hungarian laws, 1998-2006}
\format{
It is a \code{data.frame}, with 1 032 observation, 2 variables:
\describe{
\item{doc_id}{Title of the law}
\item{text}{The unprocessed texts}
}
}
\source{
\url{https://cap.tk.hu/en/dataoverview}
}
\usage{
data_lawtext_1998_2006
}
\description{
The dataset contains 1 032 Hungarian laws passed during the 1998-2002 and 2002-2006 parliamentary cycles.
}
\keyword{datasets}
