% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lawtext_1999.R
\docType{data}
\name{data_lawtext_1999}
\alias{data_lawtext_1999}
\title{Sample of Hungarian laws from 1999}
\format{
It is a \code{data.frame}, with 125 observation, 2 variables:
\describe{
\item{doc_id}{A unique document id, in this case the filename of the law text}
\item{text}{The unprocessed text}
}
}
\source{
\url{https://cap.tk.hu/en/dataoverview}
}
\usage{
data_lawtext_1999
}
\description{
The sample contains 125 Hungarian laws from 1999. It is used in the 4th chapter of the textbook (https://tankonyv.poltextlab.com/corpus-ch.html)
}
\keyword{datasets}
