% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_legal_stopwords.R
\docType{data}
\name{data_legal_stopwords}
\alias{data_legal_stopwords}
\title{Hungarian legal stopwords}
\format{
It is a character vector, with a length of 651
}
\source{
\url{tba}
}
\usage{
data_legal_stopwords
}
\description{
A collection of Hungarian legal stopwords, containing 651 stopwords. They are used in the 7th chapter of the textbook (https://tankonyv.poltextlab.com/lda-ch.html)
}
\keyword{datasets}
