% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_legal_stopwords2.R
\docType{data}
\name{data_legal_stopwords2}
\alias{data_legal_stopwords2}
\title{Additional Hungarian legal stopwords}
\format{
It is a character vector, with a length of 149.
}
\source{
\url{tba}
}
\usage{
data_legal_stopwords2
}
\description{
An additional collection of Hungarian legal stopwords, containing 149 stopwords. They should be used as an extension of the data_legal_stopwords, however we separated the two due to educational reasons. They are used in the 7th chapter of the textbook (https://tankonyv.poltextlab.com/lda-ch.html)
}
\keyword{datasets}
