% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_magyar_nemzet_small.R
\docType{data}
\name{data_magyar_nemzet_small}
\alias{data_magyar_nemzet_small}
\title{Magyar Nemzet articles (small sample)}
\format{
It is a \code{data.frame}, with 2834 observation, 3 variables:
\describe{
\item{doc_id}{A unique document id, row number in this case}
\item{text}{The unprocessed article texts}
\item{doc_date}{Date of the article}
}
}
\source{
\url{https://cap.tk.hu/en/dataoverview}
}
\usage{
data_magyar_nemzet_small
}
\description{
The dataset contains 2 834 front page articles from the print Hungarian daily, Magyar Nemzet. It is sampled from the data_magyar_nemzet_large dataset. This dataset is used in the 6th chapter of the textbook (https://tankonyv.poltextlab.com/sentiment.html).
}
\references{
Sebők, Miklós, and Zoltán Kacsuk (2021). {The Multiclass Classification of Newspaper Articles with Machine Learning: The Hybrid Binary Snowball Approach.}.
\emph{Political Analysis}, 29(2): 236-249.
}
\keyword{datasets}
