% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_miniszterelnokok.R
\docType{data}
\name{data_miniszterelnokok}
\alias{data_miniszterelnokok}
\title{Speeches of Hungarian Prime Ministers (sample)}
\format{
It is a \code{data.frame}, with 7 observation, 4 variables:
\describe{
\item{doc_id}{A unique document id}
\item{text}{The speech text}
\item{year}{Year of the speech}
\item{pm}{Name of the Prime Minister}
}
}
\source{
\url{https://cap.tk.hu/en/dataoverview}
}
\usage{
data_miniszterelnokok
}
\description{
The sample contains 7 randomly sampled speeches from the Hungarian Prime Ministers after the 1989 democraticization. It is used in the 7th chapter of the textbook (https://tankonyv.poltextlab.com/lda-ch.html).
}
\keyword{datasets}
