% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mnb_pr.R
\docType{data}
\name{data_mnb_pr}
\alias{data_mnb_pr}
\title{Central Bank of Hungary press releases}
\format{
It is a \code{data.frame}, with 180 observation, 4 variables:
\describe{
\item{date}{Date of the press release}
\item{text}{The unprocessed press release texts}
\item{id}{Unique id of the press releases}
\item{year}{Year of the press release}
}
}
\source{
\url{https://github.com/poltextlab/central_bank_communication}
}
\usage{
data_mnb_pr
}
\description{
The dataset contains 180 press releases from the Magyar Nemzeti Bank (MNB). These are released after the monetary council meetings.
}
\references{
Máté Á, Sebők M, Barczikay T (2021). \href{https://doi.org/10.1371/journal.pone.0245515}{The effect of central bank communication on sovereign bond yields: The case of Hungary}.
\emph{PLoS ONE}, 16(2): e0245515.
}
\keyword{datasets}
