% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parlspeakers_small.R
\docType{data}
\name{data_parlspeakers_small}
\alias{data_parlspeakers_small}
\title{Hungarian Parliamentary Speeches (small sample)}
\format{
It is a \code{data.frame}, with 10 observation, 4 variables:
\describe{
\item{id}{A unique document id}
\item{text}{The unprocessed speech texts}
\item{felszolalo}{Name of the speaker, Lastname Firstname (party)}
\item{part}{Party affiliation of the speaker}
}
}
\source{
\url{https://cap.tk.hu/en/dataoverview}
}
\usage{
data_parlspeakers_small
}
\description{
The dataset contains 10 speeches from Hungarian parliamentary leaders (one speech per MP). This dataset is used in the 9th chapter of the textbook (https://tankonyv.poltextlab.com/scaling.html).
}
\keyword{datasets}
