% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_petofi.R
\docType{data}
\name{data_petofi}
\alias{data_petofi}
\title{Collected works of Sándor Petőfi}
\format{
It is a \code{data.frame}, with 1 observation, 2 variables:
\describe{
\item{doc_id}{Doc id variable for easier input into quanteda}
\item{text}{The unprocessed combined works of Sándor Petőfi. All of the documents are contained as one observation (row).}
}
}
\usage{
data_petofi
}
\description{
The data contains the collected works of the famous Hungarian poet, Sándor Petőfi. It contains all of his works, without any cleaning or preprocessing.
}
\keyword{datasets}
