% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stopwords_extra.R
\docType{data}
\name{data_stopwords_extra}
\alias{data_stopwords_extra}
\title{Hungarian stopwords}
\format{
It is a character vector, with a length of 790.
}
\source{
\url{https://github.com/stopwords-iso/stopwords-hu}
}
\usage{
data_stopwords_extra
}
\description{
A collection of Hungarian stopwords, containing 790 stopwords (significantly more than the 199 in the default quanteda::stopwords("hungarian")). They are used in the 5th chapter of the textbook (https://tankonyv.poltextlab.com/leiro-stat.html)
}
\keyword{datasets}
