% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lawtext_1999.R
\docType{data}
\name{data_us_speeches}
\alias{data_us_speeches}
\title{US Speeches (May 2021)}
\format{
  It is a \code{data.frame}, with 45261 observation, 2 variables:
    \describe{
      \item{text}{Sentence-segmented speech texts}
      \item{label}{The CAP-label of the sentence}
    }
}
\source{
  \url{https://osf.io/ws5vp/}
}
\usage{
  data_us_speeches
}
\description{
  The dataset contains 45261 sentences from US speeches held in May 2021.
}
\keyword{datasets}