% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_poltext.R
\docType{data}
\name{dictionary_poltext}
\alias{dictionary_poltext}
\title{Hungarian general sentiment dictionary}
\format{
It is a quanteda Dictionary object with 2 key entries:
\describe{
\item{positive}{2 299 words with positive sentiment}
\item{negative}{2 588 words with negative sentiment}
}
}
\source{
\url{https://github.com/poltextlab/sentiment_hun}
}
\usage{
dictionary_poltext
}
\description{
This is a Hungarian general sentiment dictionary, with positive and negative dimensions. It is used in the 6th chapter in the textbook (https://tankonyv.poltextlab.com/sentiment.html).
}
\keyword{datasets}
