% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_miniszterelnokok.R
\docType{data}
\name{data_miniszterelnokok}
\alias{data_miniszterelnokok}
\title{Miniszterelnöki beszédek minta}
\format{
Egy data.frame, 7 sorral, 4 oszloppal:
\describe{
\item{doc_id}{A beszédek egyedi azonosítója}
\item{text}{A beszéd szövege}
\item{year}{A beszéd éve}
\item{pm}{A beszédet elmondó miniszterelnők neve}
}
}
\source{
\url{https://cap.tk.hu/en/dataoverview}
}
\usage{
data_miniszterelnokok
}
\description{
Az adatbázis 7 darab véletlenszerűen kiválasztott miniszterelnöki beszédet tartalmaz a rendszerváltás utáni magyar miniszterelnököktől.
}
\keyword{datasets}
